Ball moves at speed per second
The game used to have a speed per tick but the problem was when it
lagged it slowed the game down. I added Game.speedPerSecond(speed) to
the game library that changes it from speed per tick to speed per
second. Now the ball will move the same speed no matter what the fps is.